Add multiplication for posits

+++ src/aarith/posit/posit_types.hpp
@@ -8,17 +8,17 @@ namespace aarith {
 /**
  * Standard 8 bit posit with up to two exponent bits.
  */
-using posit8 = posit<8, 2, uint8_t>;
+using posit8 = posit<8, 2, uint64_t>;
 
 /**
  * Standard 16 bit posit with up to two exponent bits.
  */
-using posit16 = posit<16, 2, uint16_t>;
+using posit16 = posit<16, 2, uint64_t>;
 
 /**
  * Standard 32 bit posit with up to two exponent bits.
  */
-using posit32 = posit<32, 2, uint32_t>;
+using posit32 = posit<32, 2, uint64_t>;
 
 /**
  * Standard 64 bit posit with up to two exponent bits.
